requirements.txt
openai
python-dotenv
rich
pydantic
